+++ next-gen/src/content/links.json
@@ -1,0 +1,9 @@
+{
+  "Page": {
+    "SiteTitle": "vinckr.com",
+    "Navigation": ["index", "links"],
+    "Welcome": "hello 👋",
+    "Title": "Links",
+    "About": ["hello this is link page", "Thanks for visiting!"]
+  }
+}
