+++ next-gen/src/templates/content.tmpl
@@ -1,0 +1,16 @@
+{{define "content" -}}
+<div class="container">
+   <h2>{{.Page.Title}}</h2>
+   <h2>{{.Page.Welcome}}</h2>
+   <h3 class="title" ></h3>
+   <div class="content">  
+   {{ range .Page.About}}
+      <p>{{.}}</p>
+   {{end}}
+   <img src="./img/gif/cute.gif" alt="Greetings">
+   </div>
+</div>
+
+<div class="test">
+</div>
+{{end}}
