+++ next-gen/src/templates/footer.tmpl
