+++ next-gen/src/templates/page.tmpl
@@ -1,0 +1,5 @@
+{{define "page"}}
+{{template "header" .Page }}
+{{template "content" . }}
+{{template "footer" . }}
+{{end}}
