+++ next-gen/static/index.html
@@ -1,0 +1,44 @@
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="style.css">
+    <title>vinckr.com - Home</title>
+</head>
+<body>
+    <div class="navigation">
+            
+    <a href="./index.html">index</a>
+            
+    <a href="./links.html">links</a>
+    
+    </div>
+
+<div class="container">
+   <h2>Home</h2>
+   <h2>hello 👋</h2>
+   <h3 class="title" ></h3>
+   <div class="content">  
+   
+      <p>My name is Vincent and this is my personal website. There are many like it but this one is mine. I believe in the power and merit of open source software and a free and open internet.</p>
+   
+      <p>Feel free to write me an email to hi[ät]vinckr.com about anything, I would love to hear from you!</p>
+   
+      <p>Thanks for visiting!</p>
+   
+   <img src="./img/gif/cute.gif" alt="Greetings">
+   </div>
+</div>
+
+<div class="test">
+</div>
+
+<div class="footer">
+        test
+    </div>
+</body>
+</html>
+
