+++ next-gen/static/links.html
@@ -1,0 +1,42 @@
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="style.css">
+    <title>vinckr.com - Links</title>
+</head>
+<body>
+    <div class="navigation">
+            
+    <a href="./index.html">index</a>
+            
+    <a href="./links.html">links</a>
+    
+    </div>
+
+<div class="container">
+   <h2>Links</h2>
+   <h2>hello 👋</h2>
+   <h3 class="title" ></h3>
+   <div class="content">  
+   
+      <p>hello this is link page</p>
+   
+      <p>Thanks for visiting!</p>
+   
+   <img src="./img/gif/cute.gif" alt="Greetings">
+   </div>
+</div>
+
+<div class="test">
+</div>
+
+<div class="footer">
+        test
+    </div>
+</body>
+</html>
+
