--- next/go.mod
@@ -1,3 +1,0 @@
-module gen-index
-
-go 1.19
