--- next/src/content/index.json
@@ -1,6 +1,0 @@
-{
-  "SiteTitle": "vinckr.com",
-  "File": "index",
-  "Title": "Home",
-  "Link": "./index.html"
-}
