--- next/src/content/links.json
@@ -1,6 +1,0 @@
-{
-  "SiteTitle": "vinckr.com",
-  "File": "links",
-  "Title": "Links",
-  "Link": "./links.html"
-}
