--- next/src/templates/content.tmpl
@@ -1,28 +1,0 @@
-{{define "content" -}}
-<div class="container">
-   <h2>{{.PageTitle}}</h2>
-   <h3 class="title" >hello &#128075;,</h3>
-   <div class="content">
-      <p>
-      my name is Vincent and this is my personal website. There are many like it
-      but this one is mine.
-   </p>
-   <p>
-      I believe in the power and merit of open source software and a free and
-      open internet.
-   </p>
-   <p>
-      Feel free to write me <a href="mailto:hi@vinckr.com">an email</a> about
-      anything, I would love to hear from you!
-   </p>
-   <p>
-      Have a look around, I collect interesting <a href="/links">links</a>
-      and hack away at a <a href="/blog">blog</a>.
-   </p>
-     <img src="./img/gif/cute.gif" alt="Greetings">
-   </div>
-</div>
-
-<div class="test">
-</div>
- {{end}}
