--- next/src/templates/header.tmpl
@@ -1,21 +1,0 @@
-{{define "header" -}}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <title>{{.SiteTitle}} - {{.Title}}</title>
-    </head>
-<body>
-    <div class="navigation">
-{{range .Pages}}
-        <p><a href="{{.Link}}">{{.Title}}</a></p>
-{{end}}
-{{range $i, $a:=.}}
-        <p><a href="{{$a.Link}}">{{$a.Title}}</a></p>
-
-</div>
-{{end}}
-{{end}}
