--- next/src/templates/page.tmpl
@@ -1,5 +1,0 @@
-{{define "page"}}
-{{template "header" . -}}
-{{template "content" . -}}
-{{template "footer" -}}
-{{end}}
