--- next/static/index.html
@@ -1,15 +1,0 @@
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <title>vinckr.com - Home</title>
-    </head>
-<body>
-    <div class="navigation">
-
-
-        <p><a href="
