+++ next/welcome.html
@@ -1,0 +1,54 @@
+!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <link rel="shortcut icon" href="/favicon.png" />
+    <title>vinckr</title>
+    <link rel="canonical" href="https://vinckr.com" />
+    <meta property="og:site_name" content="vinckr" />
+    <meta property="og:title" content="vinckr" />
+    <meta property="og:url" content="https://vinckr.com" />
+    <meta property="og:type" content="website" />
+    <meta
+      property="og:description"
+      content="Hard tech seed investors supporting founders at the earliest stages. Best coffee in San Francisco ☕️"
+    />
+    <meta property="og:image" content="images/logo.png" />
+    <meta property="og:image:width" content="600" />
+    <meta property="og:image:height" content="106" />
+    <meta itemprop="name" content="vinckr" />
+    <meta itemprop="url" content="https://vinckr.com" />
+    <meta
+      itemprop="description"
+      content="Hard tech seed investors supporting founders at the earliest stages. Best coffee in San Francisco ☕️"
+    />
+    <meta itemprop="thumbnailUrl" content="images/geo/geocities.jpg" />
+    <link rel="image_src" href="images/geo/geocities.jpg" />
+    <meta itemprop="image" content="images/geo/geocities.jpg" />
+    <meta name="twitter:title" content="vinckr" />
+    <meta name="twitter:image" content="images/geo/geocities.jpg" />
+    <meta name="twitter:url" content="https://vinckr.com/welcome.htm" />
+    <meta name="twitter:card" content="summary" />
+    <meta
+      name="twitter:description"
+      content="Hard tech seed investors supporting founders at the earliest stages. Best coffee in San Francisco ☕️"
+    />
+    <meta
+      name="description"
+      content="Hard tech seed investors supporting founders at the earliest stages. Best 
+      coffee in San Francisco ☕️"
+    />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:creator" content="@rootvc" />
+    <meta name="twitter:site" content="@rootvc" />
+    <meta name="twitter:title" content="vinckr" />
+    <meta
+      name="twitter:description"
+      content="Join us! Investing in hard tech, automation, manufacturing, hardware, robotics, and more."
+    />
+    <meta name="twitter:image" content="images/geo/geocities.jpg" />
+  </head>
+</html>
